feat(#33) : model 데이터 타입 수정

+++ app/src/main/java/com/mimo/android/data/datasource/remote/MapRemoteDataSourceImpl.kt
@@ -10,9 +10,9 @@ class MapRemoteDataSourceImpl @Inject constructor(
 ) : MapRemoteDataSource {
 
     override suspend fun getMarkers(
-        latitude: Long,
-        longitude: Long,
-        radius: Long
+        latitude: Double,
+        longitude: Double,
+        radius: Double
     ): Response<MarkerResponse> {
         return mapApi.getMarkers(latitude, longitude, radius)
     }
